requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
websockets>=12.0
streamlit-audiorecorder>=0.0.5
pydub>=0.25.1

# PAPR Python SDK - install from PyPI
# For local development, the app will use ~/Documents/GitHub/papr-pythonSDK/src if available
papr-memory>=0.1.0
